// store/application/types/enums/mutation-type.ts
/**
 * Mutations supported by application store module
 */
 export enum MutationType {
    SET_IMAGE_URL_TEMPLATE = "SET_IMAGE_URL_TEMPLATE",
    SET_IMAGE_WORKAROUND_URL_TEMPLATE = "SET_IMAGE_WORKAROUND_URL_TEMPLATE",
    ADD_FOCUSED_ELEMENT = "ADD_FOCUSED_ELEMENT",
    REMOVE_FOCUSED_ELEMENT = "REMOVE_FOCUSED_ELEMENT",
    CLEAR_ALL = "CLEAR_ALL",
    OPEN_MODAL_EXIT_APP = "OPEN_MODAL_EXIT_APP",
    CLOSE_MODAL_EXIT_APP = "CLOSE_MODAL_EXIT_APP",
    SHOW_HEADER = "SHOW_HEADER",
    HIDE_HEADER = "HIDE_HEADER",
    SET_SELECTED_ID_PROGRAM = "SET_SELECTED_ID_PROGRAM",
    DELETE_SELECTED_ID_PROGRAM = "DELETE_SELECTED_ID_PROGRAM",
    SET_IS_HAVE_FOCUSED_ELEMENT = "SET_IS_HAVE_FOCUSED_ELEMENT",
    DELETE_ALL_SELECTED_ID_PROGRAM = "DELETE_ALL_SELECTED_ID_PROGRAM",
    SHOW_PREVIEW_SIGN_IN_MODAL = "SHOW_PREVIEW_SIGN_IN_MODAL",
    SET_DISABLE_ROUTE_CHANGE = "SET_DISABLE_ROUTE_CHANGE",
    SHOW_INTERNET_MODAL = "SHOW_INTERNET_MODAL",
    SET_COUNTER_VOD = "SET_COUNTER_VOD",
    SET_COUNTER_MYTV = "SET_COUNTER_MYTV",
    SET_COUNTER_SEARCH = "SET_COUNTER_SEARCH",
    SET_COUNTER_HOME = "SET_COUNTER_HOME",
    SET_COUNTER_GENRE = "SET_COUNTER_GENRE",
    SET_COUNTER_DETAIL = "SET_COUNTER_DETAIL",
    CLEAR_COUNTER_GENRE = "CLEAR_COUNTER_GENRE",
    CLEAR_COUNTER_VOD = "CLEAR_COUNTER_VOD",
    SET_TOOLTIPE_SHOWING = "SET_TOOLTIPE_SHOWING",
    SET_TOOLTIPE_MESSAGE = "SET_TOOLTIPE_MESSAGE",
    SET_PRIVACY_POLICY_TEMPLATE = "SET_PRIVACY_POLICY_TEMPLATE",
    SET_ABOUT_US_TEMPLATE = "SET_ABOUT_US_TEMPLATE",
    SET_VISIBLE_FORGOT = "SET_VISIBLE_FORGOT",
    SET_ENABLE_SENTRY = "SET_ENABLE_SENTRY",
    SET_DISABLE_SENTRY = "SET_DISABLE_SENTRY",
    SET_VISIBLE_DESCRIPTION = "SET_VISIBLE_DESCRIPTION",
    SET_CURRENT_TIME= "SET_CURRENT_TIME",
    CLEAR_COUNTER_HOME= "CLEAR_COUNTER_HOME",
    CLEAR_COUNTER_MYTV= "CLEAR_COUNTER_MYTV",
    CLEAR_COUNTER_DETAIL= "CLEAR_COUNTER_DETAIL",
    OPEN_MODAL_SIGN_OUT= "OPEN_MODAL_SIGN_OUT",
    CLOSE_MODAL_SIGN_OUT= "CLOSE_MODAL_SIGN_OUT",
    SET_FORGOT_LINK = "SET_FORGOT_LINK",
    SET_VOLUME = "SET_VOLUME",
    SET_FILTER = "SET_FILTER",
    OPEN_MODAL_EXPIRED_CODE = "OPEN_MODAL_EXPIRED_CODE",
    CLOSE_MODAL_EXPIRED_CODE = "CLOSE_MODAL_EXPIRED_CODE",
    OPEN_MODAL_SIGN_IN_UNAVAILABLE = "OPEN_MODAL_SIGN_IN_UNAVAILABLE",
    CLOSE_MODAL_SIGN_IN_UNAVAILABLE = "CLOSE_MODAL_SIGN_IN_UNAVAILABLE",
    SHOW_OKTA_TOKEN_ERROR_MODAL = "SHOW_OKTA_TOKEN_ERROR_MODAL",
    SHOW_OKTA_SERVER_ERROR_MODAL = "SHOW_OKTA_SERVER_ERROR_MODAL"
}